torch_filter/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.115.2
uvicorn[standard]==0.30.6

# PyTorch + vision (CPU wheels will be resolved by pip; use CUDA-specific index if needed)
torch
torchvision
pillow==10.4.0

# NLP stack for the text classifier
transformers==4.44.2
tokenizers==0.19.1

# Data validation
pydantic==2.9.2

python-multipart==0.0.9

httpx==0.27.2

